Complete Controller Level

+++ src/main/java/org/teamwe/carrent/controller/OrderController.java
@@ -1,12 +1,16 @@
 package org.teamwe.carrent.controller;
 
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 import org.teamwe.carrent.controller.utils.Format;
+import org.teamwe.carrent.controller.utils.SessionAttr;
 import org.teamwe.carrent.entity.Order;
+import org.teamwe.carrent.entity.User;
 import org.teamwe.carrent.service.OrderService;
 import org.teamwe.carrent.utils.ReturnStatus;
 import org.teamwe.carrent.utils.StringUtil;
 
+import javax.servlet.http.HttpSession;
 import java.util.List;
 
 /**
@@ -16,8 +20,12 @@ import java.util.List;
 
 @RestController
 public class OrderController {
+    private final OrderService service;
 
-    private OrderService service;
+    @Autowired
+    public OrderController(OrderService service) {
+        this.service = service;
+    }
 
     @PostMapping("/order")
     public Format makeOrder(@RequestParam String email,
@@ -33,7 +41,7 @@ public class OrderController {
     }
 
     @GetMapping("/user/{email}/order")
-    public Format getOrder(@PathVariable String email) {
+    public Format getOrders(@PathVariable String email) {
         if (!StringUtil.isLegalMail(email.trim())) {
             return new Format().code(ReturnStatus.FAILURE).message(StringUtil.ILLEGAL_EMAIL);
         }
@@ -48,7 +56,34 @@ public class OrderController {
     }
 
     @DeleteMapping("/order/{orderId}")
-    public Format cancelOrder(@PathVariable String orderId) {
+    public Format cancelOrder(@PathVariable String orderId, HttpSession session) {
+        Order o = service.getById(orderId);
+        if (o == null ||
+                session.getAttribute(SessionAttr.USER_ID) == null ||
+                !session.getAttribute(SessionAttr.USER_ID).equals(o.getEmail())) {
+            return new Format().code(ReturnStatus.FAILURE).message("Illegal user");
+        }
         return new Format().code(service.deleteOrder(orderId));
+    }
+
+    @DeleteMapping("/order/{id}/time")
+    public Format endOrder(@PathVariable String id) {
+        return new Format().code(service.endOrder(id));
+    }
+
+    @GetMapping("/order/{id}")
+    public Format getOrder(@PathVariable String id, HttpSession session) {
+        Order o = service.getById(id);
+        if (o == null) {
+            return new Format().code(ReturnStatus.FAILURE).message("No such of order : " + id);
+        }
+        String type = "" + session.getAttribute(SessionAttr.USER_TYPE);
+        String email = "" + session.getAttribute(SessionAttr.USER_ID);
+
+        if (String.valueOf(User.ENGENEER).equals(type)
+                || email.equals(o.getEmail())) {
+            return new Format().code(ReturnStatus.SUCCESS).addData("order", o);
+        }
+        return new Format().code(ReturnStatus.FAILURE).message("No such of order : " + id);
     }
 }
